Refactor and move import_script
